Get the plugin to load
Caveat: Some features no longer work. Fix them.

e.g. /option times you out.
